ai-worker/requirements.txt
fastapi
uvicorn
redis
langchain
langchain-openai
langchain-chroma
python-dotenv
